# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simiko API",
        "description": "API documentation for Simiko - Student Activity Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "https://simiko.rizkirmdhn.web.id/api",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/feed": {
            "get": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Get all feeds (posts and events)",
                "description": "Retrieve a list of all feeds including posts and events with optional filtering",
                "operationId": "f3d0fca51d413f454a4f508bf62fc12d",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by feed type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "post",
                                "event"
                            ]
                        }
                    },
                    {
                        "name": "ukm_id",
                        "in": "query",
                        "description": "Filter by UKM ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feed retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Feed retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FeedSummary"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/feed/{id}": {
            "get": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Get specific feed item",
                "description": "Retrieve detailed information about a specific feed item by ID",
                "operationId": "d9cd72d7891395525a60caaeff095ed2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Feed ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feed item retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Feed item retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/FeedDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Feed item not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Get posts only",
                "description": "Retrieve a list of posts only (excludes events)",
                "operationId": "4c3f4aacd6a974d6fb5600858ae109d9",
                "parameters": [
                    {
                        "name": "ukm_id",
                        "in": "query",
                        "description": "Filter by UKM ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Feed retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FeedSummary"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events": {
            "get": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Get events only",
                "description": "Retrieve a list of events only (excludes posts)",
                "operationId": "6609e0a577d5d62add8efc4f6e26a430",
                "parameters": [
                    {
                        "name": "ukm_id",
                        "in": "query",
                        "description": "Filter by UKM ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Events retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Feed retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FeedSummary"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FeedSummary": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "post",
                            "event"
                        ],
                        "example": "post"
                    },
                    "title": {
                        "type": "string",
                        "example": "Sample Post Title"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "ukm_alias": {
                        "type": "string",
                        "example": "HMIF"
                    }
                },
                "type": "object"
            },
            "FeedDetail": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "post",
                            "event"
                        ],
                        "example": "event"
                    },
                    "title": {
                        "type": "string",
                        "example": "Tech Workshop 2024"
                    },
                    "content": {
                        "type": "string",
                        "example": "Join us for an exciting tech workshop..."
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "event_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-25"
                    },
                    "event_type": {
                        "type": "string",
                        "enum": [
                            "online",
                            "offline"
                        ],
                        "example": "offline"
                    },
                    "location": {
                        "type": "string",
                        "example": "Building A, 3rd Floor"
                    },
                    "is_paid": {
                        "type": "boolean",
                        "example": true
                    },
                    "price": {
                        "type": "number",
                        "example": 50000
                    },
                    "payment_methods": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "method": {
                                    "type": "string",
                                    "example": "Bank Transfer"
                                },
                                "account_number": {
                                    "type": "string",
                                    "example": "1234567890"
                                },
                                "account_name": {
                                    "type": "string",
                                    "example": "HMIF UNS"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "ukm": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Himpunan Mahasiswa Informatika"
                            },
                            "alias": {
                                "type": "string",
                                "example": "HMIF"
                            },
                            "logo_url": {
                                "type": "string",
                                "example": "https://example.com/logo.jpg"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}